Added configuration macro
