+++ Cargo.toml
@@ -5,7 +5,6 @@ authors = ["Dylan Maccora <maccora17@gmail.com>"]
 
 [dependencies]
 pulldown-cmark = "0.0.14"
-# handlebars = "0.26"
 walkdir = "1.*"
 serde_json = "0.8.4"
 rustc-serialize = "0.*"
@@ -14,7 +13,6 @@ serde = "1.0"
 serde_yaml = "0.7"
 serde_derive = "1.0.2"
 clap = "2.24"
-# tempdir = "0.3"
 
 [dependencies.handlebars]
 version = "0.24.1"
